Add files via upload
Introduction:
If anybody is ill and wants to visit a doctor for check-up, he or she needs to visit the hospital and waits until the doctor is 
available. The patient also waits in a queue while getting appointment. If the doctor cancels the appointment for some emergency 
reasons then the patient is not able to know about the cancelation of the appointment unless or until he or she visits the hospital. As 
the mobile communication technology is developing rapidly, therefore, one can use the mobile’s applications to overcome such 
problems and inconvenience for the patients. 
Objective:
Objectives of the Project Our Modern age of technology is greatly depends on internet. Everything is converted to computer based for easier and faster communication. Here I tried to develop such a system, which will ensure some aspects, 
•	Reliability, Maintainability, cost-effectiveness and a nice user-friendly environment. 
•	The objective of the project is to provide an opportunity of getting doctor appointment easily. 
•	Proper management of appointments, securely stores all records. 
•	The main objective of the “Doctor Appointment Management System” project is to  provide easier doctor appointment and gets appointment online which save lots of  time.